cmd: migrate page command to cobra style
